Fix page transitions when reading with volume keys
